Move general scripts a level up - tests ok - but will not work in the Terraform context
Terraform will make several copies of the domain_version directory, without its peer directory context. References to the generalized scripts will not work. We need to create a true package.
Maybe it is better to do this for everything?

We still need the final "command" to be in the terraform module directory, don't we?