add submitted back without use of locals
The issue https://github.com/hashicorp/terraform/issues/16075 might have to do with the use of the rather fresh "locals" feature.
This attempts to work around that by using the timestamp() directly, without a locals.

Sadly, I cannot test this at this time. AWS is having an issue:

Amazon Route 53	Increased Propagation Time
8:21 AM PDT We are investigating increased propagation times of DNS edits to the Route 53 DNS servers. This does not impact queries to existing DNS records.
8:57 AM PDT We can confirm increased propagation times of DNS edits to the Route 53 DNS servers and continue to work towards resolution. This does not impact queries to existing DNS records.
9:38 AM PDT We have identified the root cause for the increased propagation times of DNS edits to the Route 53 DNS servers and continue to work towards resolution. This does not impact queries to existing DNS records.

As long as this is not resolved, terraform will wait for other records to 'finish' their update, before it will even try to create this TXT record.

The plan is to see
* if this change works
** if so
*** try whether a second and a third apply behave as expected
*** if so, try to add the 'ignore_changes = ['records[0]'] hack in again
**** it is not bad if the latter does not work: the service registration is confirmed with every new apply, so the submitted value _should_ change
** if not
*** change the documentation, and try to provide the submitted from outside the module with timestamp()
** if so
*** try whether a second and a third apply behave as expected
*** if that doesn't work
**** forget about the submitted for now

The ignore_changes is not supported for modules, so we cannot do that there.