fixed 'separate strings in 1 TXT record'
With all previous efforts, I either got several separate TXT records (which is not the intention of DNS-SD), or 1 TXT record with 1 long string. We need 1 TXT record with several separate strings. This commit finally realises that.
The trick is to use '\"\"' as _separator_. Together with a quote and the start and the end, and escaping along the way (of terraform and AWS), this work out.