fix a bug, update version, update documentation and keywords
I confused "at" and "submitted". "Submitted" has no place here. "At" is the time of last update, and that is always the time of last apply, i.e., timestamp. There is no "submitted", because there is no difference between the decision, and a server that accepts and executes the decision later.