fix test case
apple.com soa serial no longer is in the 'standard' format. It currently is 2010107456, which would be the 57th change 74th day of October in 2010. There is no month with 74 days.

Also aws.amazon.com and amazon.com are not in 'standard' format.

microsoft.com currently is: 2018010701, meaning 2 changes happened yesterday.

Still, the nextSoaSerial test fails.