fix issue by including the domain_name in the tag
The `DOMAIN_NAME` is included in the tag to make it possible to use this module in a configuration
more than once, for different domains. Otherwise, all uses would try to apply the same tag to the repo,
and all but the first attempts would fail. _This program does not push the tag._